fix rate varible, wrok by currency code not name

+++ app/src/main/java/com/pos/leaders/leaderspossystem/Models/Currency/Currency.java
@@ -75,7 +75,7 @@ public class Currency {
         return rate;
     }
 
-    public void setRate(long rate) {
+    public void setRate(double rate) {
         this.rate = rate;
     }
 }
